% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newcore_ini.R
\name{newcore_ini}
\alias{newcore_ini}
\title{Create a new Hector core from an INI list object}
\usage{
newcore_ini(ini_list, file = tempfile(), clean = TRUE, ...)
}
\arguments{
\item{ini_list}{Nested Hector INI configuration list object, such
as that returned by \code{\link[=read_ini]{read_ini()}}.}

\item{file}{File path for INI file. Default is \code{\link[base:tempfile]{base::tempfile()}}.}

\item{clean}{Logical. If \code{TRUE} (default), remove INI}

\item{...}{Additional arguments to \link[hector:newcore]{hector::newcore}}
}
\value{
Hector core object (see \code{\link[hector:newcore]{hector::newcore()}})
}
\description{
Simple wrapper that first writes the INI list object to a file
(\code{tempfile()} by default) and then creates a Hector core for that file.
}
